values module renamed
