+++ src/computations/mod.rs
@@ -1,13 +1,13 @@
 mod default_fns;
-mod generalized_values;
 mod heap_sort;
 mod idx;
 mod map;
+mod values;
 mod xap;
 
 pub(crate) use default_fns::*;
-pub(crate) use generalized_values::{Values, Vector, WhilstAtom, WhilstOption, WhilstVector};
 pub(crate) use heap_sort::heap_sort_into;
 pub(crate) use idx::min_opt_idx;
 pub(crate) use map::M;
+pub(crate) use values::{Values, Vector, WhilstAtom, WhilstOption, WhilstVector};
 pub(crate) use xap::X;
