+++ src/computations/values/whilst_atom.rs
@@ -1,23 +1,33 @@
 use crate::computations::{
-    Values, WhilstVector, generalized_values::whilst_iterators::WhilstOptionFlatMapIter,
+    Values, WhilstVector,
+    values::{whilst_iterators::WhilstAtomFlatMapIter, whilst_option::WhilstOption},
 };
 use orx_concurrent_bag::ConcurrentBag;
 use orx_pinned_vec::{IntoConcurrentPinnedVec, PinnedVec};
 
-pub enum WhilstOption<T> {
-    ContinueSome(T),
-    ContinueNone,
+pub enum WhilstAtom<T> {
+    Continue(T),
     Stop,
 }
 
-impl<T> Values for WhilstOption<T> {
+impl<T> WhilstAtom<T> {
+    #[inline(always)]
+    pub fn new(value: T, whilst: impl Fn(&T) -> bool) -> Self {
+        match whilst(&value) {
+            true => Self::Continue(value),
+            false => Self::Stop,
+        }
+    }
+}
+
+impl<T> Values for WhilstAtom<T> {
     type Item = T;
 
     type Error = ();
 
     fn values(self) -> impl IntoIterator<Item = Self::Item> {
         match self {
-            Self::ContinueSome(x) => Some(x).into_iter(),
+            Self::Continue(x) => Some(x).into_iter(),
             _ => None.into_iter(),
         }
     }
@@ -27,22 +37,20 @@ impl<T> Values for WhilstOption<T> {
         P: PinnedVec<Self::Item>,
     {
         match self {
-            Self::ContinueSome(x) => {
+            Self::Continue(x) => {
                 vector.push(x);
                 false
             }
-            Self::ContinueNone => false,
             Self::Stop => true,
         }
     }
 
     fn push_to_vec_with_idx(self, idx: usize, vec: &mut Vec<(usize, Self::Item)>) -> Option<usize> {
         match self {
-            Self::ContinueSome(x) => {
+            Self::Continue(x) => {
                 vec.push((idx, x));
                 None
             }
-            Self::ContinueNone => None,
             Self::Stop => Some(idx),
         }
     }
@@ -53,11 +61,10 @@ impl<T> Values for WhilstOption<T> {
         Self::Item: Send,
     {
         match self {
-            Self::ContinueSome(x) => {
+            Self::Continue(x) => {
                 bag.push(x);
                 false
             }
-            Self::ContinueNone => false,
             Self::Stop => true,
         }
     }
@@ -67,9 +74,8 @@ impl<T> Values for WhilstOption<T> {
         M: Fn(Self::Item) -> O + Clone,
     {
         match self {
-            Self::ContinueSome(x) => WhilstOption::ContinueSome(map(x)),
-            Self::ContinueNone => WhilstOption::ContinueNone,
-            Self::Stop => WhilstOption::Stop,
+            Self::Continue(x) => WhilstAtom::Continue(map(x)),
+            Self::Stop => WhilstAtom::Stop,
         }
     }
 
@@ -78,12 +84,11 @@ impl<T> Values for WhilstOption<T> {
         F: Fn(&Self::Item) -> bool + Clone,
     {
         match self {
-            Self::ContinueSome(x) => match filter(&x) {
-                true => Self::ContinueSome(x),
-                false => Self::ContinueNone,
+            Self::Continue(x) => match filter(&x) {
+                true => WhilstOption::ContinueSome(x),
+                false => WhilstOption::ContinueNone,
             },
-            Self::ContinueNone => Self::ContinueNone,
-            Self::Stop => Self::Stop,
+            Self::Stop => WhilstOption::Stop,
         }
     }
 
@@ -92,7 +97,7 @@ impl<T> Values for WhilstOption<T> {
         Vo: IntoIterator,
         Fm: Fn(Self::Item) -> Vo + Clone,
     {
-        let iter = WhilstOptionFlatMapIter::from_option(self, &flat_map);
+        let iter = WhilstAtomFlatMapIter::from_atom(self, &flat_map);
         WhilstVector(iter)
     }
 
@@ -101,26 +106,11 @@ impl<T> Values for WhilstOption<T> {
         Fm: Fn(Self::Item) -> Option<O>,
     {
         match self {
-            Self::ContinueSome(x) => match filter_map(x) {
+            Self::Continue(x) => match filter_map(x) {
                 Some(x) => WhilstOption::ContinueSome(x),
                 None => WhilstOption::ContinueNone,
             },
-            Self::ContinueNone => WhilstOption::ContinueNone,
             Self::Stop => WhilstOption::Stop,
-        }
-    }
-
-    fn whilst(self, whilst: impl Fn(&Self::Item) -> bool) -> impl Values<Item = Self::Item>
-    where
-        Self: Sized,
-    {
-        match self {
-            Self::ContinueSome(x) => match whilst(&x) {
-                true => Self::ContinueSome(x),
-                false => Self::Stop,
-            },
-            Self::ContinueNone => Self::ContinueNone,
-            Self::Stop => Self::Stop,
         }
     }
 
@@ -129,12 +119,24 @@ impl<T> Values for WhilstOption<T> {
         X: Fn(Self::Item, Self::Item) -> Self::Item,
     {
         match self {
-            Self::ContinueSome(x) => match acc {
+            Self::Continue(x) => match acc {
                 Some(acc) => (false, Some(reduce(acc, x))),
                 None => (false, Some(x)),
             },
-            Self::ContinueNone => (false, acc),
             Self::Stop => (true, acc),
+        }
+    }
+
+    fn whilst(self, whilst: impl Fn(&Self::Item) -> bool) -> impl Values<Item = Self::Item>
+    where
+        Self: Sized,
+    {
+        match self {
+            Self::Continue(x) => match whilst(&x) {
+                true => Self::Continue(x),
+                false => Self::Stop,
+            },
+            Self::Stop => Self::Stop,
         }
     }
 
@@ -143,16 +145,18 @@ impl<T> Values for WhilstOption<T> {
         X: Fn(&mut U, Self::Item, Self::Item) -> Self::Item,
     {
         match self {
-            Self::ContinueSome(x) => match acc {
+            Self::Continue(x) => match acc {
                 Some(acc) => Some(reduce(u, acc, x)),
                 None => Some(x),
             },
-            Self::ContinueNone => acc,
             Self::Stop => acc,
         }
     }
 
     fn first(self) -> WhilstOption<Self::Item> {
-        self
+        match self {
+            Self::Continue(x) => WhilstOption::ContinueSome(x),
+            Self::Stop => WhilstOption::Stop,
+        }
     }
 }
