+++ src/computations/values/whilst_option.rs
@@ -1,33 +1,23 @@
 use crate::computations::{
-    Values, WhilstVector,
-    generalized_values::{whilst_iterators::WhilstAtomFlatMapIter, whilst_option::WhilstOption},
+    Values, WhilstVector, values::whilst_iterators::WhilstOptionFlatMapIter,
 };
 use orx_concurrent_bag::ConcurrentBag;
 use orx_pinned_vec::{IntoConcurrentPinnedVec, PinnedVec};
 
-pub enum WhilstAtom<T> {
-    Continue(T),
+pub enum WhilstOption<T> {
+    ContinueSome(T),
+    ContinueNone,
     Stop,
 }
 
-impl<T> WhilstAtom<T> {
-    #[inline(always)]
-    pub fn new(value: T, whilst: impl Fn(&T) -> bool) -> Self {
-        match whilst(&value) {
-            true => Self::Continue(value),
-            false => Self::Stop,
-        }
-    }
-}
-
-impl<T> Values for WhilstAtom<T> {
+impl<T> Values for WhilstOption<T> {
     type Item = T;
 
     type Error = ();
 
     fn values(self) -> impl IntoIterator<Item = Self::Item> {
         match self {
-            Self::Continue(x) => Some(x).into_iter(),
+            Self::ContinueSome(x) => Some(x).into_iter(),
             _ => None.into_iter(),
         }
     }
@@ -37,20 +27,22 @@ impl<T> Values for WhilstAtom<T> {
         P: PinnedVec<Self::Item>,
     {
         match self {
-            Self::Continue(x) => {
+            Self::ContinueSome(x) => {
                 vector.push(x);
                 false
             }
+            Self::ContinueNone => false,
             Self::Stop => true,
         }
     }
 
     fn push_to_vec_with_idx(self, idx: usize, vec: &mut Vec<(usize, Self::Item)>) -> Option<usize> {
         match self {
-            Self::Continue(x) => {
+            Self::ContinueSome(x) => {
                 vec.push((idx, x));
                 None
             }
+            Self::ContinueNone => None,
             Self::Stop => Some(idx),
         }
     }
@@ -61,10 +53,11 @@ impl<T> Values for WhilstAtom<T> {
         Self::Item: Send,
     {
         match self {
-            Self::Continue(x) => {
+            Self::ContinueSome(x) => {
                 bag.push(x);
                 false
             }
+            Self::ContinueNone => false,
             Self::Stop => true,
         }
     }
@@ -74,8 +67,9 @@ impl<T> Values for WhilstAtom<T> {
         M: Fn(Self::Item) -> O + Clone,
     {
         match self {
-            Self::Continue(x) => WhilstAtom::Continue(map(x)),
-            Self::Stop => WhilstAtom::Stop,
+            Self::ContinueSome(x) => WhilstOption::ContinueSome(map(x)),
+            Self::ContinueNone => WhilstOption::ContinueNone,
+            Self::Stop => WhilstOption::Stop,
         }
     }
 
@@ -84,11 +78,12 @@ impl<T> Values for WhilstAtom<T> {
         F: Fn(&Self::Item) -> bool + Clone,
     {
         match self {
-            Self::Continue(x) => match filter(&x) {
-                true => WhilstOption::ContinueSome(x),
-                false => WhilstOption::ContinueNone,
+            Self::ContinueSome(x) => match filter(&x) {
+                true => Self::ContinueSome(x),
+                false => Self::ContinueNone,
             },
-            Self::Stop => WhilstOption::Stop,
+            Self::ContinueNone => Self::ContinueNone,
+            Self::Stop => Self::Stop,
         }
     }
 
@@ -97,7 +92,7 @@ impl<T> Values for WhilstAtom<T> {
         Vo: IntoIterator,
         Fm: Fn(Self::Item) -> Vo + Clone,
     {
-        let iter = WhilstAtomFlatMapIter::from_atom(self, &flat_map);
+        let iter = WhilstOptionFlatMapIter::from_option(self, &flat_map);
         WhilstVector(iter)
     }
 
@@ -106,24 +101,12 @@ impl<T> Values for WhilstAtom<T> {
         Fm: Fn(Self::Item) -> Option<O>,
     {
         match self {
-            Self::Continue(x) => match filter_map(x) {
+            Self::ContinueSome(x) => match filter_map(x) {
                 Some(x) => WhilstOption::ContinueSome(x),
                 None => WhilstOption::ContinueNone,
             },
+            Self::ContinueNone => WhilstOption::ContinueNone,
             Self::Stop => WhilstOption::Stop,
-        }
-    }
-
-    fn acc_reduce<X>(self, acc: Option<Self::Item>, reduce: X) -> (bool, Option<Self::Item>)
-    where
-        X: Fn(Self::Item, Self::Item) -> Self::Item,
-    {
-        match self {
-            Self::Continue(x) => match acc {
-                Some(acc) => (false, Some(reduce(acc, x))),
-                None => (false, Some(x)),
-            },
-            Self::Stop => (true, acc),
         }
     }
 
@@ -132,11 +115,26 @@ impl<T> Values for WhilstAtom<T> {
         Self: Sized,
     {
         match self {
-            Self::Continue(x) => match whilst(&x) {
-                true => Self::Continue(x),
+            Self::ContinueSome(x) => match whilst(&x) {
+                true => Self::ContinueSome(x),
                 false => Self::Stop,
             },
+            Self::ContinueNone => Self::ContinueNone,
             Self::Stop => Self::Stop,
+        }
+    }
+
+    fn acc_reduce<X>(self, acc: Option<Self::Item>, reduce: X) -> (bool, Option<Self::Item>)
+    where
+        X: Fn(Self::Item, Self::Item) -> Self::Item,
+    {
+        match self {
+            Self::ContinueSome(x) => match acc {
+                Some(acc) => (false, Some(reduce(acc, x))),
+                None => (false, Some(x)),
+            },
+            Self::ContinueNone => (false, acc),
+            Self::Stop => (true, acc),
         }
     }
 
@@ -145,18 +143,16 @@ impl<T> Values for WhilstAtom<T> {
         X: Fn(&mut U, Self::Item, Self::Item) -> Self::Item,
     {
         match self {
-            Self::Continue(x) => match acc {
+            Self::ContinueSome(x) => match acc {
                 Some(acc) => Some(reduce(u, acc, x)),
                 None => Some(x),
             },
+            Self::ContinueNone => acc,
             Self::Stop => acc,
         }
     }
 
     fn first(self) -> WhilstOption<Self::Item> {
-        match self {
-            Self::Continue(x) => WhilstOption::ContinueSome(x),
-            Self::Stop => WhilstOption::Stop,
-        }
+        self
     }
 }
